IBM Logo Rom functional
draw instruction implemented

+++ core/src/main/java/dev/bartel/chip8/Instruction.java
@@ -2,5 +2,5 @@ package dev.bartel.chip8;
 
 @FunctionalInterface
 public interface Instruction {
-    public void execute(Cpu cpu, short opcode);
+    public void execute(Cpu cpu, int opcode);
 }
